# src/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple CRUD API for managing vehicle information.",
        "title": "Vehicle API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Lekan Adetunmbi",
            "url": "https://pollaroid.net",
            "email": "me@pollaroid.net"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "api-vehicle.onrender.com",
    "basePath": "/v1",
    "paths": {
...
        "/makes": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get makes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "returns all makes",
                "responses": {
                    "200": {
                        "description": "all makes returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Make"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "create make",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "creates a new make",
                "parameters": [
                    {
                        "description": "Make Details JSON",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateMakeDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Make"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request body validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "409": {
                        "description": "make with the same name exists",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        },
        "/makes/countries": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get makes by country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "returns makes by country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "make search by country",
                        "name": "country",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Make"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request query validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "no make with the search country found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        },
        "/makes/names": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get make by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "returns a make by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "make search by name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "make with the search name returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Make"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request query validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "no make with the search name found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        },
        "/makes/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get make by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "returns a make by its 16 caharcter uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Make ID(UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "make with the specified ID returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Make"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request param validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "make with the specified ID not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "delete make",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "deletes a make",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Make ID(UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "make deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.SuccessResponseDto"
                        }
                    },
                    "400": {
                        "description": "request params validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "make with ID in request params not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "update make",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Make"
                ],
                "summary": "updates a make",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Make ID(UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Make Details JSON",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateMakeDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "make updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Make"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request body/param validation errors or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "make with ID in request params not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "409": {
                        "description": "another make with the same name in request body exists",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.CreateMakeDto": {
            "type": "object",
            "required": [
                "country",
                "name"
            ],
            "properties": {
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dtos.FailedResponseDto": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "errorType": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
        "dtos.SuccessResponseDto": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Make": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vehicle"
                    }
                }
            }
        },
...
        "models.Vehicle": {
            "type": "object",
            "properties": {
                "acceleration": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "cylinder": {
                    "type": "integer"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "horsePower": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "make": {
                    "$ref": "#/definitions/models.Make"
                },
                "makeID": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "numberOfSeats": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "topSpeed": {
                    "type": "string"
                },
                "torque": {
                    "type": "string"
                },
                "transmission": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}